requirements.txt

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-multipart==0.0.6

# Database
pymongo==4.6.0
mongoengine==0.27.0

# Data & Analytics
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Market Data APIs
polygon-api-client==1.12.4
requests==2.31.0

# Trading & Execution
quickfix==1.15.1

# Task Scheduling
celery==5.3.4
redis==5.0.1

# Data Storage
pyarrow==14.0.1
boto3==1.33.6

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
python-dateutil==2.8.2
pytz==2023.3

# Monitoring & Logging
structlog==23.2.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
black==23.11.0
flake8==6.1.0
